refactor: Extract namespace CLI tests to their own module
